working/agent/requirements.txt
fastapi>=0.115.0
mangum>=0.19.0
uvicorn>=0.32.0
pydantic>=2.0.0
boto3>=1.35.0
pymysql>=1.1.0
cryptography>=41.0.0
langgraph>=0.2.0
langchain-aws>=0.2.0
langchain-google-genai>=2.0.0
langchain-core>=0.3.0
langchain-community>=0.3.0
httpx>=0.27.0
python-dateutil>=2.9.0
uuid6>=2024.1.12
PyJWT>=2.8.0
python-telegram-bot>=21.0
apscheduler>=3.10.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
python-dotenv>=1.0.0
pypdf>=3.0.0
tiktoken>=0.7.0
google-generativeai>=0.8.0
